+++ app/src/main/kotlin/dev/skymansandy/gocorona/presentation/statedata/StateDataFragment.kt
@@ -50,24 +50,24 @@ class StateDataFragment(override val layoutId: Int = R.layout.fragment_state_dat
                 binding.layoutStatList.setup(covidRes, statAdapter, getString(R.string.district))
 
                 with(binding.statsNonIndia) {
-                    tvActiveCount.showNumber(newState.active.toInt())
-                    tvConfirmedCount.showNumber(newState.confirmed.toInt())
-                    tvRecoveredCount.showNumber(newState.recovered.toInt())
-                    tvDeceasedCount.showNumber(newState.deaths.toInt())
+                    tvActiveCount.showNumber(newState.active)
+                    tvConfirmedCount.showNumber(newState.confirmed)
+                    tvRecoveredCount.showNumber(newState.recovered)
+                    tvDeceasedCount.showNumber(newState.deaths)
 
                     if (tvRecoveredCount.scanForBigTextAndWrapNextLine())
                         tvDeceasedDelta.setOrientation(LinearLayout.VERTICAL)
                     if (tvDeceasedCount.scanForBigTextAndWrapNextLine())
                         tvRecoveredDelta.setOrientation(LinearLayout.VERTICAL)
 
-                    showDelta(covidRes, tvConfirmedDelta, newState.confirmedToday.toInt())
-                    showDelta(covidRes, tvRecoveredDelta, newState.recoveredToday.toInt())
-                    showDelta(covidRes, tvDeceasedDelta, newState.deathsToday.toInt())
+                    showDelta(covidRes, tvConfirmedDelta, newState.confirmedToday)
+                    showDelta(covidRes, tvRecoveredDelta, newState.recoveredToday)
+                    showDelta(covidRes, tvDeceasedDelta, newState.deathsToday)
                     pieChart.loadData(
                         covidRes,
-                        newState.active.toInt(),
-                        newState.recovered.toInt(),
-                        newState.deaths.toInt()
+                        newState.active,
+                        newState.recovered,
+                        newState.deaths
                     )
                 }
                 binding.layoutStatList.root.visibility = if (newState.stats.isNullOrEmpty()) {
